+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Once we reach the v1.0 release, this project will adhere to [Semantic Versioning
 ## [Unreleased]
 - Use error wrapping from standard library instead of `github.com/pgk/errors`
 - Update Module to Go 1.14
+- Change default log level from Debug to Info
 
 ## [v0.10.0] - 2019-10-26
 - Allow event handlers to also use scalar event types (fixes #14)